Add HoldTime value back in
Following discussion with Jeff, Claudio, Ben Maddison, and Ben Cox,
we observed that SendHoldTimer depends on at least one message being
send within every SendHoldTime period. If the negotiated HoldTime is
zero, no KEEPALIVEs will be send within the HoldTime period of time,
and thus the SendHoldTimer is likely to expire.

This adding HoldTime back in as this was a regression.